refactor(update_arena): Splitted up all methods and structs in separate files
