+++ replay_parser/src/packet_parser/events/entity_method/avatar_methods/mod.rs
@@ -1,0 +1,1 @@
+pub mod update_arena;
