+++ replay_parser/src/packet_parser/events/entity_method/avatar_methods/update_arena/avatar_ready.rs
@@ -1,0 +1,26 @@
+use serde_pickle::Value as PickleVal;
+
+use super::UpdateData;
+use crate::packet_parser::prelude::*;
+
+#[derive(Debug, Clone, Serialize)]
+pub struct AvatarReady {
+    vehicle_id: i32,
+}
+
+pub fn parse_avatar_ready(arena_data: &[u8]) -> Result<UpdateData, PacketError> {
+    let pickle_value = serde_pickle::value_from_slice(
+        arena_data,
+        serde_pickle::DeOptions::new().replace_unresolved_globals(),
+    )?;
+
+    let PickleVal::I64(vehicle_id) = pickle_value  else {
+        return Err(PacketError::PickleError("Expected I64 value for vehicle id".into()))
+    };
+
+    let avatar_ready = AvatarReady {
+        vehicle_id: vehicle_id as i32,
+    };
+
+    Ok(UpdateData::AvatarReady(avatar_ready))
+}
