+++ replay_parser/src/packet_parser/events/entity_method/mod.rs
@@ -1,9 +1,9 @@
-mod update_arena;
+mod avatar_methods;
 pub(crate) mod vehicle_methods;
 // mod vehicle_misc_status;
 
+use avatar_methods::update_arena::UpdateArena;
 use nom::number::complete::le_i32;
-use update_arena::UpdateArena;
 pub use vehicle_methods::*;
 
 use crate::packet_parser::prelude::*;
